Move stream support into main file, drop support for extending node primitives
